Permission applied to Category & User apis

+++ app/Http/Controllers/API/V1/CategoryController.php
@@ -4,19 +4,22 @@ namespace App\Http\Controllers\API\V1;
 
 use App\Http\Requests\Categories\CategoryStoreRequest;
 use App\Http\Requests\Categories\CategoryUpdateRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Category;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Tools\Response;
+use App\Tools\Permission;
 use Illuminate\Support\Facades\DB;
 
 class CategoryController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize(Permission::format("view", 'category'), Category::class);
         return Response::success('Categories found.', ['categories'=> Category::paginate(10)]);
     }
 
@@ -25,6 +28,7 @@ class CategoryController extends Controller
      */
     public function store(CategoryStoreRequest $request)
     {
+        $this->authorize(Permission::format("create", 'category'), Category::class);
         $category = new Category();
 
         DB::transaction(function () use ($request, & $category) {
@@ -42,6 +46,7 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
+        $this->authorize(Permission::format("view", 'category'), Category::class);
         return Response::success('Category found.', ['category'=> $category]);
     }
 
@@ -50,6 +55,7 @@ class CategoryController extends Controller
      */
     public function update(CategoryUpdateRequest $request, Category $category)
     {
+        $this->authorize(Permission::format("update", 'category'), Category::class);
         if($request->title == $category->title && $request->status == $category->status)
         {
             return Response::success('Nothing to update.');
@@ -69,6 +75,7 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize(Permission::format("delete", 'category'), Category::class);
         DB::transaction(function () use ($category) {
             $category->delete();
         });
