+++ app/Http/Controllers/API/V1/UserController.php
@@ -8,14 +8,18 @@ use App\Http\Requests\Users\UserUpdateRequest;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use App\Tools\Response;
+use App\Tools\Permission;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize(Permission::format("view", 'user'), User::class);
         return Response::success('Users data found.', ['users'=> User::paginate(10)]);
     }
 
@@ -24,6 +28,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize(Permission::format("create", 'user'), User::class);
         return Response::success('Please use register endpoint here "/api/v1/register"');
     }
 
@@ -32,6 +37,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize(Permission::format("view", 'user'), User::class);
         return Response::success('User data found.', ['user'=> $user]);
     }
 
@@ -40,6 +46,7 @@ class UserController extends Controller
      */
     public function update(UserUpdateRequest $request, User $user)
     {
+        $this->authorize(Permission::format("update", 'user'), User::class);
         if($request->role == $user->role && $request->status == $user->status)
         {
             return Response::success('Nothing to update.');
@@ -59,6 +66,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize(Permission::format("delete", 'user'), User::class);
         DB::transaction(function () use ($user) {
             $user->delete();
         });
